Formatted code.

+++ src/thermal/thermal.cpp
@@ -4,41 +4,53 @@
 
 #include "thermal.hpp"
 
-bool exists(const std::string& name){
-    struct stat buffer;   
-    return (stat (name.c_str(), &buffer) == 0);
+bool exists(const std::string &name)
+{
+    struct stat buffer;
+    return (stat(name.c_str(), &buffer) == 0);
 }
 
-ThermalReader::ThermalReader(std::string zone0, std::string zone1){
-    if(exists(zone0)){
+ThermalReader::ThermalReader(std::string zone0, std::string zone1)
+{
+    if (exists(zone0))
+    {
         path_zone0 = zone0;
-    } else {
+    }
+    else
+    {
         throw "File for zone0 doesn't exists";
     }
 
-    if(exists(zone1)){
+    if (exists(zone1))
+    {
         path_zone1 = zone1;
-    } else {
+    }
+    else
+    {
         throw "File for zone1 doesn't exists";
     }
 }
 
-int get_temperature_from_file(const std::string& path){
+int get_temperature_from_file(const std::string &path)
+{
     std::string line;
-    std::ifstream file (path);
-    if(file.good()){
-        getline(file,line);
+    std::ifstream file(path);
+    if (file.good())
+    {
+        getline(file, line);
         return std::stoi(line);
     }
     return 0;
 }
 
-int ThermalReader::get_temperature(){
+double ThermalReader::get_temperature()
+{
     int thermal0 = get_temperature_from_file(path_zone0);
     int thermal1 = get_temperature_from_file(path_zone1);
 
-    if(thermal0 > thermal1){
-        return thermal0/1000;
+    if (thermal0 > thermal1)
+    {
+        return thermal0 / 1000.0;
     }
-    return thermal1/1000;
+    return thermal1 / 1000.0;
 }
