+++ src/thermal/thermal.hpp
@@ -1,10 +1,11 @@
 #include <string>
 
-class ThermalReader{
+class ThermalReader
+{
     std::string path_zone0;
     std::string path_zone1;
 
-    public:
-        ThermalReader(std::string path_zone0, std::string path_zone1);
-        int get_temperature();
+public:
+    ThermalReader(std::string path_zone0, std::string path_zone1);
+    double get_temperature();
 };
